+++ Yousician.srv/src/config/app.php
@@ -50,6 +50,7 @@ return [
         \App\Providers\AuthServiceProvider::class,
         \App\Providers\EventServiceProvider::class,
         \App\Providers\DatabaseServiceProvider::class,
+        \App\Providers\BladeServiceProvider::class,
     ],
     'aliases' => [
         'Auth' => \App\Support\Auth::class,