+++ Yousician.srv/src/resources/views/sections/navigation/logo.blade.php
@@ -1,0 +1,3 @@
+<a href='/' id="logo">
+    Slim
+</a>