+++ Yousician.srv/src/resources/views/sections/navigation/top-left.blade.php
@@ -1,0 +1,7 @@
+@include('sections.navigation.logo')
+
+@if (Auth::check())
+    <a href="/home">
+        Home
+    </a>
+@endif