+++ Yousician.srv/src/resources/views/sections/navigation/top-right.blade.php
@@ -1,0 +1,12 @@
+@if (Auth::guest())
+    <a href="/register">
+        Register
+    </a>
+    <a href="/login">
+        Login
+    </a>
+@else
+    <a href="/logout">
+        Logout
+    </a>
+@endif