+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### 1.0.3 (2024-06-15)
+SERVER:
+- fix upload backtrack
+- add blade template provider
+- add functional for third party service
+-- add youtube upload backtrack
+
 ### 1.0.2 (2023-11-27)
 APP:
 - fix button sigle note text color
@@ -9,7 +16,5 @@ APP:
 - remove browscap
 - add sillabus switch instrument route and method
 
-
 ### 1.0.0 (2023-10-30)
-
 - initial commit